#392 Rename squbs-streamingpipeline to squbs-pipeline
