+++ docs/actor-hierarchy.md
@@ -13,7 +13,7 @@ squbs sets up the actor and component hierarchy shown in the above picture to su
 
 * **RouteDefinition/FlowDefinition** - `RouteDefinition` and `FlowDefinition` are both service definitions of different forms. A `RouteDefinition` defines the `Route` for a service whereas a `FlowDefinition` defines a `Flow`. These two are different types of request handlers handling incoming Http requests. Both `RouteDefinition` and `FlowDefinition` themselves are not actors, but a class extending their respective traits. They are registered by their respective services and are hosted by their corresponding actors.
 
-* **Route/Flow actors** - Actors of type `org.squbs.unicomplex.streaming.RouteActor` or `org.squbs.unicomplex.streaming.FlowActor` created by the `CubeSupervisor` to host each corresponding `RouteDefinition` or `FlowDefinition`. Therefore, they become children supervised by the corresponding `CubeSupervisor` they belong to. Their request handlers are automatically registered against all the listeners they are bound to, allowing requests from the listeners to be dispatched to the corresponding request handlers.
+* **Route/Flow actors** - Actors of type `org.squbs.unicomplex.RouteActor` or `org.squbs.unicomplex.FlowActor` created by the `CubeSupervisor` to host each corresponding `RouteDefinition` or `FlowDefinition`. Therefore, they become children supervised by the corresponding `CubeSupervisor` they belong to. Their request handlers are automatically registered against all the listeners they are bound to, allowing requests from the listeners to be dispatched to the corresponding request handlers.
 
 * **Request handler actors** - Instead of using `RouteDefinition`/`FlowDefinition` for services, developers may choose to work at the low level API and thus register a request handler actor instead. Avoiding the high level routing API generally consumes less memory and allows handling of streaming requests but is harder to code when compared to the routing DSL.
