+++ squbs-httpclient/src/test/scala/org/squbs/httpclient/ClientFlowPipelineSpec.scala
@@ -25,8 +25,8 @@ import akka.stream.{ActorMaterializer, BidiShape}
 import akka.util.ByteString
 import com.typesafe.config.ConfigFactory
 import org.scalatest.{AsyncFlatSpec, BeforeAndAfterAll, Matchers}
+import org.squbs.pipeline.{Context, PipelineFlow, PipelineFlowFactory, RequestContext}
 import org.squbs.resolver.ResolverRegistry
-import org.squbs.pipeline.streaming._
 import org.squbs.testkit.Timeouts._
 
 import scala.concurrent.{Await, Future}
