+++ squbs-pipeline/src/main/scala/org/squbs/pipeline/PipelineExtension.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.squbs.pipeline.streaming
+package org.squbs.pipeline
 
 import akka.NotUsed
 import akka.actor._
